+++ demisto_sdk/commands/download/downloader.py
@@ -131,7 +131,8 @@ class Downloader:
         :return: True if fetched successfully, False otherwise
         """
         try:
-            self.client = demisto_client.configure(verify_ssl=not self.insecure)
+            verify = (not self.insecure) if self.insecure else None  # set to None so demisto_client will use env var DEMISTO_VERIFY_SSL
+            self.client = demisto_client.configure(verify_ssl=verify)
             api_response: tuple = demisto_client.generic_request_func(self.client, '/content/bundle', 'GET')
             body: bytes = ast.literal_eval(api_response[0])
             io_bytes = io.BytesIO(body)
@@ -214,10 +215,9 @@ class Downloader:
         """
         output_pack_path = self.output_pack_path
         if not (os.path.isdir(output_pack_path) and
-                os.path.basename(os.path.dirname(os.path.abspath(output_pack_path))) == 'Packs' and
-                os.path.basename(os.path.dirname(os.path.dirname(os.path.abspath(output_pack_path)))) == 'content'):
+                os.path.basename(os.path.dirname(os.path.abspath(output_pack_path))) == 'Packs'):
             print_color(f"Path {output_pack_path} is not a valid Path pack. The designated output pack's path is"
-                        f" of format ~/.../content/Packs/$PACK_NAME", LOG_COLORS.RED)
+                        f" of format ~/.../Packs/$PACK_NAME", LOG_COLORS.RED)
             return False
         return True