+++ demisto_sdk/commands/format/format_module.py
@@ -8,6 +8,7 @@ from demisto_sdk.commands.common.tools import (find_type, get_files_in_dir,
 from demisto_sdk.commands.format.format_constants import SCHEMAS_PATH
 from demisto_sdk.commands.format.update_classifier import (
     ClassifierJSONFormat, OldClassifierJSONFormat)
+from demisto_sdk.commands.format.update_connection import ConnectionJSONFormat
 from demisto_sdk.commands.format.update_dashboard import DashboardJSONFormat
 from demisto_sdk.commands.format.update_incidentfields import \
     IncidentFieldJSONFormat
@@ -46,9 +47,9 @@ FILE_TYPE_AND_LINKED_CLASS = {
     'pythonfile': PythonFileFormat,
     'report': ReportJSONFormat,
     'testscript': ScriptYMLFormat,
+    'canvas-context-connections': ConnectionJSONFormat
 }
-UNFORMATTED_FILES = ['canvas-context-connections',
-                     'readme',
+UNFORMATTED_FILES = ['readme',
                      'releasenotes',
                      'description',
                      'changelog',
@@ -62,8 +63,13 @@ VALIDATE_RES_SKIPPED_CODE = 2
 VALIDATE_RES_FAILED_CODE = 3
 
 
-def format_manager(input: str = None, output: str = None, from_version: str = '', no_validate: bool = False,
-                   verbose: bool = False, update_docker: bool = False):
+def format_manager(input: str = None,
+                   output: str = None,
+                   from_version: str = '',
+                   no_validate: bool = False,
+                   verbose: bool = False,
+                   update_docker: bool = False,
+                   assume_yes: bool = False):
     """
     Format_manager is a function that activated format command on different type of files.
     Args:
@@ -73,6 +79,7 @@ def format_manager(input: str = None, output: str = None, from_version: str = ''
         no_validate (flag): Whether the user specifies not to run validate after format.
         verbose (bool): Whether to print verbose logs or not
         update_docker (flag): Whether to update the docker image.
+        assume_yes (bool): Whether to assume "yes" as answer to all prompts and run non-interactively
     Returns:
         int 0 in case of success 1 otherwise
     """
@@ -93,9 +100,14 @@ def format_manager(input: str = None, output: str = None, from_version: str = ''
             file_type = find_type(file_path)
             if file_type and file_type.value not in UNFORMATTED_FILES:
                 file_type = file_type.value
-                info_res, err_res, skip_res = run_format_on_file(input=file_path, file_type=file_type,
-                                                                 from_version=from_version, output=output,
-                                                                 no_validate=no_validate, verbose=verbose, update_docker=update_docker)
+                info_res, err_res, skip_res = run_format_on_file(input=file_path,
+                                                                 file_type=file_type,
+                                                                 from_version=from_version,
+                                                                 output=output,
+                                                                 no_validate=no_validate,
+                                                                 verbose=verbose,
+                                                                 update_docker=update_docker,
+                                                                 assume_yes=assume_yes)
                 if err_res:
                     error_list.append("err_res")
                 if err_res: