+++ demisto_sdk/commands/format/update_playbook.py
@@ -14,10 +14,16 @@ from demisto_sdk.commands.format.update_generic_yml import BaseUpdateYML
 
 
 class BasePlaybookYMLFormat(BaseUpdateYML):
-    def __init__(self, input: str = '', output: str = '', path: str = '', from_version: str = '',
-                 no_validate: bool = False, verbose: bool = False):
+    def __init__(self,
+                 input: str = '',
+                 output: str = '',
+                 path: str = '',
+                 from_version: str = '',
+                 no_validate: bool = False,
+                 verbose: bool = False,
+                 assume_yes: bool = False):
         super().__init__(input=input, output=output, path=path, from_version=from_version, no_validate=no_validate,
-                         verbose=verbose)
+                         verbose=verbose, assume_yes=assume_yes)
 
     def add_description(self):
         """Add empty description to playbook and tasks."""