+++ demisto_sdk/commands/lint/helpers.py
@@ -425,13 +425,17 @@ def split_warnings_errors(output: str):
             list of error messags, list of warnings messages, list of all undetected messages
         """
     output_lst = output.split('\n')
+    # Warnings and errors lists currently relevant for XSOAR Linter
     warnings_list = []
     error_list = []
+    # Others list is relevant for mypy and flake8.
     other_msg_list = []
     for msg in output_lst:
-        if msg.startswith('W') or 'W90' in msg:
+        # 'W:' for python2 xsoar linter
+        # 'W[0-9]' for python3 xsoar linter
+        if (msg.startswith('W') and msg[1].isdigit()) or 'W:' in msg or 'W90' in msg:
             warnings_list.append(msg)
-        elif msg.startswith('E') or 'E90' in msg:
+        elif (msg.startswith('E') and msg[1].isdigit()) or 'E:' in msg or 'E90' in msg:
             error_list.append(msg)
         else:
             other_msg_list.append(msg)