+++ demisto_sdk/tests/integration_tests/format_integration_test.py
@@ -127,6 +127,73 @@ def test_integration_format_yml_with_no_test_negative(tmp_path: PosixPath,
     os.remove(saved_file_path)
 
 
+@pytest.mark.parametrize('source_path,destination_path,formatter,yml_title,file_type', BASIC_YML_TEST_PACKS)
+def test_integration_format_yml_with_no_test_no_interactive_positive(tmp_path: PosixPath,
+                                                                     source_path: str,
+                                                                     destination_path: str,
+                                                                     formatter: BaseUpdateYML,
+                                                                     yml_title: str,
+                                                                     file_type: str):
+    """
+        Given
+        - A yml file (integration, playbook or script) with no 'tests' configured
+
+        When
+        - using the '-y' option
+
+        Then
+        -  Ensure no exception is raised
+        -  Ensure 'No tests' is added in the first time
+    """
+    saved_file_path = str(tmp_path / os.path.basename(destination_path))
+    runner = CliRunner()
+    # Running format in the first time
+    result = runner.invoke(main, [FORMAT_CMD, '-i', source_path, '-o', saved_file_path, '-y'])
+    assert not result.exception
+    yml_content = get_dict_from_file(saved_file_path)
+    assert yml_content[0].get('tests') == ['No tests (auto formatted)']
+
+
+@pytest.mark.parametrize('source_path,destination_path,formatter,yml_title,file_type', YML_FILES_WITH_TEST_PLAYBOOKS)
+def test_integration_format_configuring_conf_json_no_interactive_positive(tmp_path: PosixPath,
+                                                                          source_path: str,
+                                                                          destination_path: str,
+                                                                          formatter: BaseUpdateYML,
+                                                                          yml_title: str,
+                                                                          file_type: str):
+    """
+        Given
+        - A yml file (integration, playbook or script) with no tests playbooks configured that are not configured
+            in conf.json
+
+        When
+        - using the -y option
+
+        Then
+        -  Ensure no exception is raised
+        -  If file_type is playbook or a script: Ensure {"playbookID": <content item ID>} is added to conf.json
+            for each test playbook configured in the yml under 'tests' key
+        -  If file_type is integration: Ensure {"playbookID": <content item ID>, "integrations": yml_title} is
+            added to conf.json for each test playbook configured in the yml under 'tests' key
+    """
+    # Setting up conf.json
+    conf_json_path = str(tmp_path / 'conf.json')
+    with open(conf_json_path, 'w') as file:
+        json.dump(CONF_JSON_ORIGINAL_CONTENT, file, indent=4)
+    BaseUpdateYML.CONF_PATH = conf_json_path
+
+    test_playbooks = ['test1', 'test2']
+    saved_file_path = str(tmp_path / os.path.basename(destination_path))
+    runner = CliRunner()
+    # Running format in the first time
+    result = runner.invoke(main, [FORMAT_CMD, '-i', source_path, '-o', saved_file_path, '-y'])
+    assert not result.exception
+    if file_type == 'playbook':
+        _verify_conf_json_modified(test_playbooks, '', conf_json_path)
+    else:
+        _verify_conf_json_modified(test_playbooks, yml_title, conf_json_path)
+
+
 @pytest.mark.parametrize('source_path,destination_path,formatter,yml_title,file_type', YML_FILES_WITH_TEST_PLAYBOOKS)
 def test_integration_format_configuring_conf_json_positive(tmp_path: PosixPath,
                                                            source_path: str,